+++ extensions/cornerstone/src/handleSegmentationStorage.js
@@ -1,10 +1,10 @@
-import * as dcmjs from "dcmjs";
+import * as dcmjs from 'dcmjs';
 
-import OHIF from "@ohif/core";
-import cornerstone from "cornerstone-core";
-import cornerstoneTools from "cornerstone-tools";
+import OHIF from '@ohif/core';
+import cornerstone from 'cornerstone-core';
+import cornerstoneTools from 'cornerstone-tools';
 
-const { StackManager } = OHIF.utils;
+const { StackManager, DicomLoaderService } = OHIF.utils;
 
 function getDisplaySet(studies, studyInstanceUid, displaySetInstanceUid) {
   const study = studies.find(
@@ -52,14 +52,6 @@ function addSegMetadataToCornerstoneToolState(
   }
 }
 
-function retrieveDicomData(wadoUri) {
-  // TODO: Authorization header depends on the server. If we ever have multiple servers
-  // we will need to figure out how / when to pass this information in.
-  return fetch(wadoUri, {
-    headers: OHIF.DICOMWeb.getAuthorizationHeader()
-  }).then(response => response.arrayBuffer());
-}
-
 async function handleSegmentationStorage(
   studies,
   studyInstanceUid,
@@ -73,8 +65,11 @@ async function handleSegmentationStorage(
     studyInstanceUid,
     displaySetInstanceUid
   );
-  const segWadoUri = displaySet.images[0].getData().wadouri;
-  const arrayBuffer = await retrieveDicomData(segWadoUri);
+
+  const arrayBuffer = await DicomLoaderService.findDicomDataPromise(
+    displaySet,
+    studies
+  );
   const dicomData = dcmjs.data.DicomMessage.readFile(arrayBuffer);
   const dataset = dcmjs.data.DicomMetaDictionary.naturalizeDataset(
     dicomData.dict
@@ -91,7 +86,7 @@ async function handleSegmentationStorage(
 
   if (displaySets.length > 1) {
     console.warn(
-      "More than one display set with the same seriesInstanceUid. This is not supported yet..."
+      'More than one display set with the same seriesInstanceUid. This is not supported yet...'
     );
   }
 
@@ -100,17 +95,19 @@ async function handleSegmentationStorage(
   const results = parseSeg(arrayBuffer, imageIds);
 
   if (!results) {
-    throw new Error("Fractional segmentations are not supported");
+    throw new Error('Fractional segmentations are not supported');
   }
 
-  const { segMetadata, toolState } = results;
+  const { labelmapBuffer, segMetadata, segmentsOnFrame } = results;
+  const { setters } = cornerstoneTools.getModule('segmentation');
 
-  segMetadata.seriesInstanceUid = seriesInstanceUid;
-
-  addSegMetadataToCornerstoneToolState(
+  setters.labelmap3DByFirstImageId(
+    imageIds[0],
+    labelmapBuffer,
+    0, // TODO -> Can define a color LUT based on colors in the SEG later.
     segMetadata,
-    toolState,
-    displaySetInstanceUid
+    imageIds.length,
+    segmentsOnFrame
   );
 
   const cachedStack = StackManager.findOrCreateStack(
@@ -118,12 +115,13 @@ async function handleSegmentationStorage(
     referenceDisplaySet
   );
   const stack = Object.assign({}, cachedStack);
+
   stack.currentImageIdIndex = 0;
 
   return {
     studyInstanceUid,
     displaySetInstanceUid,
-    stack
+    stack,
   };
 }