+++ extensions/dicom-pdf/src/OHIFDicomPDFViewport.js
@@ -1,86 +1,59 @@
-import React, { Component } from "react";
-import PropTypes from "prop-types";
-import OHIF from "@ohif/core";
-import OHIFComponentPlugin from "./OHIFComponentPlugin.js";
-import DicomPDFViewport from "./DicomPDFViewport";
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import OHIF from '@ohif/core';
+import OHIFComponentPlugin from './OHIFComponentPlugin.js';
+import DicomPDFViewport from './DicomPDFViewport';
 
-const { DICOMWeb } = OHIF;
+const { DicomLoaderService } = OHIF.utils;
 
 class OHIFDicomPDFViewport extends Component {
   static propTypes = {
     studies: PropTypes.object,
     displaySet: PropTypes.object,
-    viewportIndex: PropTypes.number
+    viewportIndex: PropTypes.number,
   };
 
   state = {
     byteArray: null,
-    error: null
+    error: null,
   };
 
-  static id = "DicomPDFViewportPDF";
+  static id = 'DicomPDFViewportPDF';
 
   static init() {
-    console.log("DicomPDFViewport init()");
+    console.log('DicomPDFViewport init()');
   }
 
   static destroy() {
-    console.log("DicomPDFViewport destroy()");
+    console.log('DicomPDFViewport destroy()');
   }
 
   componentDidMount() {
-    const { displaySet } = this.props.viewportData;
+    const { displaySet, studies } = this.props.viewportData;
     const {
       studyInstanceUid,
       seriesInstanceUid,
       sopInstanceUid,
       wadoRoot,
       wadoUri,
-      authorizationHeaders
+      authorizationHeaders,
     } = displaySet;
 
-    this.retrieveDicomData(
-      studyInstanceUid,
-      seriesInstanceUid,
-      sopInstanceUid,
-      wadoRoot,
-      wadoUri,
-      authorizationHeaders
-    ).then(
-      byteArray => {
+    DicomLoaderService.findDicomDataPromise(displaySet, studies).then(
+      data => {
+        const byteArray = new Uint8Array(data);
         this.setState({
-          byteArray
+          byteArray: byteArray,
         });
       },
       error => {
         this.setState({
-          error
+          error,
         });
 
         throw new Error(error);
       }
     );
-  }
-
-  retrieveDicomData(
-    studyInstanceUid,
-    seriesInstanceUid,
-    sopInstanceUid,
-    wadoRoot,
-    wadoUri,
-    authorizationHeaders
-  ) {
-    // TODO: Passing in a lot of data we aren't using
-
-    // TODO: Authorization header depends on the server. If we ever have multiple servers
-    // we will need to figure out how / when to pass this information in.
-    return fetch(wadoUri, {
-      headers: authorizationHeaders
-    })
-      .then(response => response.arrayBuffer())
-      .then(arraybuffer => {
-        return new Uint8Array(arraybuffer);
-      });
   }
 
   render() {