+++ platform/core/src/DICOMWeb/getAuthorizationHeader.js
@@ -6,17 +6,26 @@ import user from '../user';
 /**
  * Returns the Authorization header as part of an Object.
  *
- * @returns {Object}
+ * @export
+ * @param {Object} [server={}]
+ * @param {Object} [server.requestOptions]
+ * @param {string|function} [server.requestOptions.auth]
+ * @returns {Object} { Authorization }
  */
-export default function getAuthorizationHeader(server) {
+export default function getAuthorizationHeader({ requestOptions } = {}) {
   const headers = {};
 
   // Check for OHIF.user since this can also be run on the server
   const accessToken = user && user.getAccessToken && user.getAccessToken();
 
-  if (server && server.requestOptions && server.requestOptions.auth) {
-    // HTTP Basic Auth (user:password)
-    headers.Authorization = `Basic ${btoa(server.requestOptions.auth)}`;
+  if (requestOptions && requestOptions.auth) {
+    if (typeof requestOptions.auth === 'function') {
+      // Custom Auth Header
+      headers.Authorization = requestOptions.auth(requestOptions);
+    } else {
+      // HTTP Basic Auth (user:password)
+      headers.Authorization = `Basic ${btoa(requestOptions.auth)}`;
+    }
   } else if (accessToken) {
     headers.Authorization = `Bearer ${accessToken}`;
   }