+++ platform/viewer/public/config/demo.js
@@ -12,9 +12,7 @@ window.config = {
         qidoSupportsIncludeField: true,
         imageRendering: 'wadors',
         thumbnailRendering: 'wadors',
-        requestOptions: {
-          requestFromBrowser: true,
-        },
+        enableStudyLazyLoad: true,
       },
     ],
   },