+++ platform/viewer/public/config/netlify.js
@@ -11,9 +11,7 @@ window.config = {
         qidoSupportsIncludeField: true,
         imageRendering: 'wadors',
         thumbnailRendering: 'wadors',
-        requestOptions: {
-          requestFromBrowser: true,
-        },
+        enableStudyLazyLoad: true,
       },
     ],
   },